refac: em disciplinas
