+++ src/app.js
@@ -1,4 +1,5 @@
 import professores from "./controllers/Professores.js";
+import disciplina from "./controllers/Disciplina.js";
 import express from "express";
 import cors from "cors";
 
@@ -6,7 +7,7 @@ const app = express();
 
 app.use(cors());
 app.use(express.json());
-app.use(professores);
+app.use(professores, disciplina);
 
 //Quando não encontrar alguma rota
 app.use((req,res,next) => {
